Influx-Injector: Add data consumption from UNS
