+++ influxdb-injector/.env.example
@@ -1,0 +1,8 @@
+INFLUX_URL=http://localhost:8086/
+INFLUX_TOKEN=WLUvPm_7qJ7vV4ARpR9v86sbEUFEGXS_YPCxSLYhWjbq_hRLGrGMxHqeyzMmVPLIGzWNUpfRbhXCXwp5PxhwSA==
+INFLUX_ORG=default
+INFLUX_BUCKET=default
+LOG_LEVEL=info
+BATCH_SIZE=5000
+FLUSH_INTERVAL=10000
+MQTT_URL=mqtt.local.acs.app.amrc.co.uk:1883
