+++ influxdb-injector/package.json
@@ -4,6 +4,8 @@
   "description": "An ingestion engine to enable InfluxDB to be a fully-featured Sparkplug historian",
   "author": "AMRC",
   "license": "MIT",
+  "main": "./src/app.js",
+  "type": "module",
   "keywords": [
     "amrc",
     "factory-plus",
@@ -13,22 +15,20 @@
     "mqtt",
     "sparkplug"
   ],
-  "type": "module",
   "scripts": {
-    "start": "node --trace-warnings dist/bin/ingester.js",
-    "start:shell": "k5start -Uf $CLIENT_KEYTAB npm run start",
-    "dev": " npm run build && npm run start:dev",
-    "build": "tsc && tsc-alias",
+    "start": "node dist/app.js",
+    "dev": " npm run build && npm run start",
+    "build": "tsc",
     "clean": "tsc --build --clean",
     "lint": "eslint --ignore-path .gitignore --ext .ts src/",
     "lint:fix": "npm run lint -- --fix"
   },
   "dependencies": {
-    "@amrc-factoryplus/utilities": "^1.0.8",
     "@influxdata/influxdb-client": "^1.33.2",
     "@influxdata/influxdb-client-apis": "^1.33.2",
     "async": "^3.2.4",
     "long": "^5.2.3",
+    "mqtt": "^5.9.1",
     "pino": "^8.11.0",
     "pino-pretty": "^10.0.0"
   },
