+++ influxdb-injector/run-dev.sh
@@ -4,7 +4,4 @@
 #
 
 # bin/bash
-tmpfile=$(mktemp)
-export CLIENT_KEYTAB="$(kubectl --kubeconfig ~/.kube/fplus get -n factory-plus secret krb5-keytabs -o jsonpath="{.data.sv1warehouse}" | base64 -d >"$tmpfile" && echo "$tmpfile")"
-echo $CLIENT_KEYTAB
-npm run start:shell
+npm run start
